Backend/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
django-cors-headers
python-dotenv
requests
Pillow
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2